requirements.txt
requests
coqui-tts
pydub
pyserial
bleak